Extract SDK types and remove register pins

+++ include/psyq/snd.h
@@ -139,8 +139,8 @@ long SsUtPitchBend(long voice, long vab_id, long program, long note, u_short pbe
 long SsUtChangePitch(long voice, long vab_id, long program, long old_note, long old_fine, long new_note, long new_fine);
 long SsUtChangeADSR(long voice, long vab_id, long program, long old_note, long adsr1, long adsr2);
 long SsUtGetDetVVol(long voice, short *left, short *right);
-long SsUtSetDetVVol(long voice, short left, short right);
-long SsUtSetVVol(long voice, short left, short right);
+short SsUtSetDetVVol(short voice, short left, short right);
+short SsUtSetVVol(short voice, short left, short right);
 long SsUtAutoVol(long voice, long start_vol, long end_vol, long delta_time);
 long SsUtAutoPan(long voice, long start_pan, long end_pan, long delta_time);
 void SsSeqSetNoteParam2C(long seq, long sep, u_char value);
